add User object to auth hook
